core: Make time argument to internal foo_notify_bar functions 64 bit
The libinput evdev code uses 64 bit timestamps internally, to avoid having to
deal with timestamp wraps. The internal foo_notify_bar functions time argument
however is only 32 bits, bump this to 64 bits to avoid truncating the timestamps
when calling these functions.

This is a preparation patch for adding internal event listeners, so that the
callbacks for these can get the full 64 bit timestamps.

+++ src/libinput-private.h
@@ -213,37 +213,37 @@ notify_removed_device(struct libinput_device *device);
 
 void
 keyboard_notify_key(struct libinput_device *device,
-		    uint32_t time,
+		    uint64_t time,
 		    uint32_t key,
 		    enum libinput_key_state state);
 
 void
 pointer_notify_motion(struct libinput_device *device,
-		      uint32_t time,
+		      uint64_t time,
 		      double dx,
 		      double dy);
 
 void
 pointer_notify_motion_absolute(struct libinput_device *device,
-			       uint32_t time,
+			       uint64_t time,
 			       double x,
 			       double y);
 
 void
 pointer_notify_button(struct libinput_device *device,
-		      uint32_t time,
+		      uint64_t time,
 		      int32_t button,
 		      enum libinput_button_state state);
 
 void
 pointer_notify_axis(struct libinput_device *device,
-		    uint32_t time,
+		    uint64_t time,
 		    enum libinput_pointer_axis axis,
 		    double value);
 
 void
 touch_notify_touch_down(struct libinput_device *device,
-			uint32_t time,
+			uint64_t time,
 			int32_t slot,
 			int32_t seat_slot,
 			double x,
@@ -251,7 +251,7 @@ touch_notify_touch_down(struct libinput_device *device,
 
 void
 touch_notify_touch_motion(struct libinput_device *device,
-			  uint32_t time,
+			  uint64_t time,
 			  int32_t slot,
 			  int32_t seat_slot,
 			  double x,
@@ -259,13 +259,13 @@ touch_notify_touch_motion(struct libinput_device *device,
 
 void
 touch_notify_touch_up(struct libinput_device *device,
-		      uint32_t time,
+		      uint64_t time,
 		      int32_t slot,
 		      int32_t seat_slot);
 
 void
 touch_notify_frame(struct libinput_device *device,
-		   uint32_t time);
+		   uint64_t time);
 
 static inline uint64_t
 libinput_now(struct libinput *libinput)
